# Use shared_ptr to store cse_intermediate_results in LambdaDoubleVisitor