- this class contains lambdas which capture the address of cse_intermediate_results
  - so a copy of this class refers to the same address
  - shared_ptr ensures the address remains valid when original LambdaDoubleVisitor goes out of scope
  - test case copied from https://github.com/symengine/symengine/pull/1673
  - resolves #1674