@isuruf 2020-08-27
Thanks for sending the PR. Wouldn't this give wrong results if the two visitors are used simultaneously in two threads?