@lkeegan 2020-08-27
That's a good point - an alternative that I think would be thread safe (if a bit uglier) would be to pass a pointer to the `cse_intermediate_results` vector as a function parameter instead of capturing it implicitly in the lambdas