@lkeegan 2020-08-28
I tried benchmarking (#1682) this PR and passing the pointer everywhere does add some measurable overhead (about 4% for this example). There is probably a better way to fix this issue without making it slower..

Benchmark without this PR:
```
benchmarking lambda_real_double_visitor_call
collecting 10000 samples, 1 iterations each, in estimated 744.6 ms
mean: 662.56 ns, lb 662.093 ns, ub 663.3 ns, ci 0.95
std dev: 29.456 ns, lb 20.8268 ns, ub 43.2118 ns, ci 0.95
found 28 outliers among 10000 samples (0.28%)
variance is severely inflated by outliers

benchmarking lambda_real_double_visitor_cse_call
collecting 10000 samples, 1 iterations each, in estimated 1131.2 ms
mean: 253.112 ns, lb 252.43 ns, ub 255.449 ns, ci 0.95
std dev: 56.8014 ns, lb 14.3531 ns, ub 128.943 ns, ci 0.95
found 609 outliers among 10000 samples (6.09%)
variance is severely inflated by outliers
```
with this PR:
```
benchmarking lambda_real_double_visitor_call
collecting 10000 samples, 1 iterations each, in estimated 773.91 ms
mean: 682.057 ns, lb 681.665 ns, ub 683.063 ns, ci 0.95
std dev: 29.6895 ns, lb 14.2783 ns, ub 60.7058 ns, ci 0.95
found 30 outliers among 10000 samples (0.3%)
variance is severely inflated by outliers

benchmarking lambda_real_double_visitor_cse_call
collecting 10000 samples, 1 iterations each, in estimated 1120.5 ms
mean: 263.931 ns, lb 263.183 ns, ub 266.967 ns, ci 0.95
std dev: 67.2083 ns, lb 14.0266 ns, ub 161.427 ns, ci 0.95
found 32 outliers among 10000 samples (0.32%)
variance is severely inflated by outliers
```