@isuruf 2020-08-28
Thanks @lkeegan for the benchmarks. @lkeegan, @bjodah, do you have a use-case for using the copy constructor? (Can we delete the copy constructor so that people don't get surprised?) This seems to be slower and also break where these classes are extended externally. (For eg: @jppelteret extends this class in deal.II)