@bjodah 2020-08-28
I'm fine with deleting the copy constructor (I had a visitor as a member in a class, upon copying that outer class I encountered this bug, I wrapped it in a shared_ptr to avoid it being copied when copying my encapsulating class).