@lkeegan 2020-08-28
I'd also be fine with deleting the copy constructor. I'm not sure if the move constructor would have the same issue? I guess if the vector is move constructed the address of the elements shouldn't change.